moving pool_rdm to avoid circular import
